+++ src/bad/robot/refactoring/chapter1/RegularPrice.java
@@ -3,8 +3,8 @@ package bad.robot.refactoring.chapter1;
 import java.math.BigDecimal;
 
 public class RegularPrice extends Price {
-    private static final double DAILY_LATE_FEE = 1.5;
-    private static final int BASE_CHARGE = 2;
+    private static final BigDecimal DAILY_LATE_FEE = BigDecimal.valueOf(1.5);
+    private static final BigDecimal BASE_CHARGE = BigDecimal.valueOf(2);
     private static final int RENTAL_PERIOD = 2;
 
     @Override
